I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

The QueriesQueueManager relies on Meta Semaphore, which needs to schedule multiple queries per second. I noticed that the time between when a query releases the semaphore and when the next query acquires it is too long and inconsistent.

To address this, I refactored the semaphore release process so that it no longer checks for permit drops during extension, but instead performs the release immediately after the permits are dropped.

---

use https://github.com/databendlabs/wizard/tree/main/threads to benchmark.
Enable in config:
```
global_statement_queue = true
max_running_queries = 8
```

Before:
```
python bench.py --runbend --total 2000 --threads 32 --case select
Benchmarking completed in 241.86 seconds.

==================================================
📊 BENCHMARK SUMMARY - CASE: SELECT - DATABEND
==================================================
• Operations: 2000 completed in 241.85s
• Concurrency: 32 threads (97.1% utilized)

📈 PERFORMANCE METRICS:
• Time per operation: 0.12s/op (total time / operations = 241.85s/2000)
• Throughput: 8.27 ops/s (avg), 27.94 ops/s (peak)
• Efficiency: 97.5% of theoretical maximum

📊 THROUGHPUT DISTRIBUTION:
0-10 ops/s: ██████████ (55%)
10-25 ops/s: ████████ (44%)
25-50 ops/s:  (1%)
50-75 ops/s:  (0%)
75-100 ops/s:  (0%)
==================================================
```

After:
```
python bench.py --runbend --total 2000 --threads 32 --case select
==================================================
📊 BENCHMARK SUMMARY - CASE: SELECT - DATABEND
==================================================
• Operations: 2000 completed in 44.30s
• Concurrency: 32 threads (95.5% utilized)

📈 PERFORMANCE METRICS:
• Time per operation: 0.02s/op (total time / operations = 44.30s/2000)
• Throughput: 45.15 ops/s (avg), 73.70 ops/s (peak)
• Efficiency: 96.8% of theoretical maximum

📊 THROUGHPUT DISTRIBUTION:
0-10 ops/s:  (2%)
10-25 ops/s: █ (7%)
25-50 ops/s: ██████████ (53%)
50-75 ops/s: ███████ (38%)
75-100 ops/s:  (0%)
==================================================
```



## Tests

- [ ] Unit Test
- [ ] Logic Test
- [x] Benchmark Test
- [ ] No Test - _Explain why_

## Type of change

- [ ] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [x] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/databendlabs/databend/17999)
<!-- Reviewable:end -->